+++ src/packages/property-editors/checkbox-list/components/input-checkbox-list/index.ts
